+++ sdk/resourcemanager/datafactory/armdatafactory/models.go
@@ -7770,6 +7770,69 @@ type CreateRunResponse struct {
 	RunID *string `json:"runId,omitempty"`
 }
 
+// CredentialClassification provides polymorphic access to related types.
+// Call the interface's GetCredential() method to access the common type.
+// Use a type switch to determine the concrete type.  The possible types are:
+// - *Credential, *ManagedIdentityCredential, *ServicePrincipalCredential
+type CredentialClassification interface {
+	// GetCredential returns the Credential content of the underlying type.
+	GetCredential() *Credential
+}
+
+// Credential - The Azure Data Factory nested object which contains the information and credential which can be used to connect
+// with related store or compute resource.
+type Credential struct {
+	// REQUIRED; Type of credential.
+	Type *string `json:"type,omitempty"`
+
+	// OPTIONAL; Contains additional key/value pairs not defined in the schema.
+	AdditionalProperties map[string]interface{}
+
+	// List of tags that can be used for describing the Credential.
+	Annotations []interface{} `json:"annotations,omitempty"`
+
+	// Credential description.
+	Description *string `json:"description,omitempty"`
+}
+
+// GetCredential implements the CredentialClassification interface for type Credential.
+func (c *Credential) GetCredential() *Credential { return c }
+
+// CredentialListResponse - A list of credential resources.
+type CredentialListResponse struct {
+	// REQUIRED; List of credentials.
+	Value []*ManagedIdentityCredentialResource `json:"value,omitempty"`
+
+	// The link to the next page of results, if any remaining results exist.
+	NextLink *string `json:"nextLink,omitempty"`
+}
+
+// CredentialOperationsClientCreateOrUpdateOptions contains the optional parameters for the CredentialOperationsClient.CreateOrUpdate
+// method.
+type CredentialOperationsClientCreateOrUpdateOptions struct {
+	// ETag of the credential entity. Should only be specified for update, for which it should match existing entity or can be
+	// * for unconditional update.
+	IfMatch *string
+}
+
+// CredentialOperationsClientDeleteOptions contains the optional parameters for the CredentialOperationsClient.Delete method.
+type CredentialOperationsClientDeleteOptions struct {
+	// placeholder for future optional parameters
+}
+
+// CredentialOperationsClientGetOptions contains the optional parameters for the CredentialOperationsClient.Get method.
+type CredentialOperationsClientGetOptions struct {
+	// ETag of the credential entity. Should only be specified for get. If the ETag matches the existing entity tag, or if * was
+	// provided, then no content will be returned.
+	IfNoneMatch *string
+}
+
+// CredentialOperationsClientListByFactoryOptions contains the optional parameters for the CredentialOperationsClient.ListByFactory
+// method.
+type CredentialOperationsClientListByFactoryOptions struct {
+	// placeholder for future optional parameters
+}
+
 // CredentialReference - Credential reference type.
 type CredentialReference struct {
 	// REQUIRED; Reference credential name.
@@ -16357,14 +16420,14 @@ func (l *LinkedIntegrationRuntimeType) GetLinkedIntegrationRuntimeType() *Linked
 // - *JiraLinkedService, *LinkedService, *MagentoLinkedService, *MariaDBLinkedService, *MarketoLinkedService, *MicrosoftAccessLinkedService,
 // - *MongoDbAtlasLinkedService, *MongoDbLinkedService, *MongoDbV2LinkedService, *MySQLLinkedService, *NetezzaLinkedService,
 // - *ODataLinkedService, *OdbcLinkedService, *Office365LinkedService, *OracleCloudStorageLinkedService, *OracleLinkedService,
-// - *OracleServiceCloudLinkedService, *PaypalLinkedService, *PhoenixLinkedService, *PostgreSQLLinkedService, *PrestoLinkedService,
-// - *QuickBooksLinkedService, *QuickbaseLinkedService, *ResponsysLinkedService, *RestServiceLinkedService, *SQLServerLinkedService,
-// - *SalesforceLinkedService, *SalesforceMarketingCloudLinkedService, *SalesforceServiceCloudLinkedService, *SapBWLinkedService,
-// - *SapCloudForCustomerLinkedService, *SapEccLinkedService, *SapHanaLinkedService, *SapOdpLinkedService, *SapOpenHubLinkedService,
-// - *SapTableLinkedService, *ServiceNowLinkedService, *SftpServerLinkedService, *SharePointOnlineListLinkedService, *ShopifyLinkedService,
-// - *SmartsheetLinkedService, *SnowflakeLinkedService, *SparkLinkedService, *SquareLinkedService, *SybaseLinkedService, *TeamDeskLinkedService,
-// - *TeradataLinkedService, *TwilioLinkedService, *VerticaLinkedService, *WebLinkedService, *XeroLinkedService, *ZendeskLinkedService,
-// - *ZohoLinkedService
+// - *OracleServiceCloudLinkedService, *PaypalLinkedService, *PhoenixLinkedService, *PostgreSQLLinkedService, *PowerBILinkedService,
+// - *PrestoLinkedService, *QuickBooksLinkedService, *QuickbaseLinkedService, *ResponsysLinkedService, *RestServiceLinkedService,
+// - *SQLServerLinkedService, *SalesforceLinkedService, *SalesforceMarketingCloudLinkedService, *SalesforceServiceCloudLinkedService,
+// - *SapBWLinkedService, *SapCloudForCustomerLinkedService, *SapEccLinkedService, *SapHanaLinkedService, *SapOdpLinkedService,
+// - *SapOpenHubLinkedService, *SapTableLinkedService, *ServiceNowLinkedService, *SftpServerLinkedService, *SharePointOnlineListLinkedService,
+// - *ShopifyLinkedService, *SmartsheetLinkedService, *SnowflakeLinkedService, *SparkLinkedService, *SquareLinkedService,
+// - *SybaseLinkedService, *TeamDeskLinkedService, *TeradataLinkedService, *TwilioLinkedService, *VerticaLinkedService, *WebLinkedService,
+// - *XeroLinkedService, *ZendeskLinkedService, *ZohoLinkedService
 type LinkedServiceClassification interface {
 	// GetLinkedService returns the LinkedService content of the underlying type.
 	GetLinkedService() *LinkedService
@@ -16738,6 +16801,58 @@ func (m *MagentoSource) GetTabularSource() *TabularSource {
 		DisableMetricsCollection: m.DisableMetricsCollection,
 		AdditionalProperties:     m.AdditionalProperties,
 	}
+}
+
+// ManagedIdentityCredential - Managed identity credential.
+type ManagedIdentityCredential struct {
+	// REQUIRED; Type of credential.
+	Type *string `json:"type,omitempty"`
+
+	// OPTIONAL; Contains additional key/value pairs not defined in the schema.
+	AdditionalProperties map[string]interface{}
+
+	// List of tags that can be used for describing the Credential.
+	Annotations []interface{} `json:"annotations,omitempty"`
+
+	// Credential description.
+	Description *string `json:"description,omitempty"`
+
+	// Managed identity credential properties.
+	TypeProperties *ManagedIdentityTypeProperties `json:"typeProperties,omitempty"`
+}
+
+// GetCredential implements the CredentialClassification interface for type ManagedIdentityCredential.
+func (m *ManagedIdentityCredential) GetCredential() *Credential {
+	return &Credential{
+		Type:                 m.Type,
+		Description:          m.Description,
+		Annotations:          m.Annotations,
+		AdditionalProperties: m.AdditionalProperties,
+	}
+}
+
+// ManagedIdentityCredentialResource - Credential resource type.
+type ManagedIdentityCredentialResource struct {
+	// REQUIRED; Managed Identity Credential properties.
+	Properties *ManagedIdentityCredential `json:"properties,omitempty"`
+
+	// READ-ONLY; Etag identifies change in the resource.
+	Etag *string `json:"etag,omitempty" azure:"ro"`
+
+	// READ-ONLY; The resource identifier.
+	ID *string `json:"id,omitempty" azure:"ro"`
+
+	// READ-ONLY; The resource name.
+	Name *string `json:"name,omitempty" azure:"ro"`
+
+	// READ-ONLY; The resource type.
+	Type *string `json:"type,omitempty" azure:"ro"`
+}
+
+// ManagedIdentityTypeProperties - Managed identity type properties.
+type ManagedIdentityTypeProperties struct {
+	// The resource id of user assigned managed identity
+	ResourceID *string `json:"resourceId,omitempty"`
 }
 
 // ManagedIntegrationRuntime - Managed integration runtime, including managed elastic and managed dedicated integration runtimes.
@@ -21012,6 +21127,51 @@ type PostgreSQLTableDatasetTypeProperties struct {
 
 	// This property will be retired. Please consider using schema + table properties instead.
 	TableName interface{} `json:"tableName,omitempty"`
+}
+
+// PowerBILinkedService - Power BI linked service.
+type PowerBILinkedService struct {
+	// REQUIRED; Type of linked service.
+	Type *string `json:"type,omitempty"`
+
+	// REQUIRED; Power BI linked service properties.
+	TypeProperties *PowerBILinkedServiceTypeProperties `json:"typeProperties,omitempty"`
+
+	// OPTIONAL; Contains additional key/value pairs not defined in the schema.
+	AdditionalProperties map[string]interface{}
+
+	// List of tags that can be used for describing the linked service.
+	Annotations []interface{} `json:"annotations,omitempty"`
+
+	// The integration runtime reference.
+	ConnectVia *IntegrationRuntimeReference `json:"connectVia,omitempty"`
+
+	// Linked service description.
+	Description *string `json:"description,omitempty"`
+
+	// Parameters for linked service.
+	Parameters map[string]*ParameterSpecification `json:"parameters,omitempty"`
+}
+
+// GetLinkedService implements the LinkedServiceClassification interface for type PowerBILinkedService.
+func (p *PowerBILinkedService) GetLinkedService() *LinkedService {
+	return &LinkedService{
+		Type:                 p.Type,
+		ConnectVia:           p.ConnectVia,
+		Description:          p.Description,
+		Parameters:           p.Parameters,
+		Annotations:          p.Annotations,
+		AdditionalProperties: p.AdditionalProperties,
+	}
+}
+
+// PowerBILinkedServiceTypeProperties - Power BI linked service properties.
+type PowerBILinkedServiceTypeProperties struct {
+	// REQUIRED; The tenant id to which the Power BI workspace belongs.
+	TenantID *string `json:"tenantId,omitempty"`
+
+	// REQUIRED; The Power BI workspace id.
+	WorkspaceID *string `json:"workspaceId,omitempty"`
 }
 
 // PowerQuerySink - Power query sink.
@@ -25979,6 +26139,46 @@ func (s *ServiceNowSource) GetTabularSource() *TabularSource {
 	}
 }
 
+// ServicePrincipalCredential - Service principal credential.
+type ServicePrincipalCredential struct {
+	// REQUIRED; Type of credential.
+	Type *string `json:"type,omitempty"`
+
+	// REQUIRED; Service Principal credential properties.
+	TypeProperties *ServicePrincipalCredentialTypeProperties `json:"typeProperties,omitempty"`
+
+	// OPTIONAL; Contains additional key/value pairs not defined in the schema.
+	AdditionalProperties map[string]interface{}
+
+	// List of tags that can be used for describing the Credential.
+	Annotations []interface{} `json:"annotations,omitempty"`
+
+	// Credential description.
+	Description *string `json:"description,omitempty"`
+}
+
+// GetCredential implements the CredentialClassification interface for type ServicePrincipalCredential.
+func (s *ServicePrincipalCredential) GetCredential() *Credential {
+	return &Credential{
+		Type:                 s.Type,
+		Description:          s.Description,
+		Annotations:          s.Annotations,
+		AdditionalProperties: s.AdditionalProperties,
+	}
+}
+
+// ServicePrincipalCredentialTypeProperties - Service Principal credential type properties.
+type ServicePrincipalCredentialTypeProperties struct {
+	// The app ID of the service principal used to authenticate
+	ServicePrincipalID interface{} `json:"servicePrincipalId,omitempty"`
+
+	// The key of the service principal used to authenticate.
+	ServicePrincipalKey *AzureKeyVaultSecretReference `json:"servicePrincipalKey,omitempty"`
+
+	// The ID of the tenant to which the service principal belongs
+	Tenant interface{} `json:"tenant,omitempty"`
+}
+
 // SetVariableActivity - Set value for a Variable.
 type SetVariableActivity struct {
 	// REQUIRED; Activity name.
@@ -26823,6 +27023,9 @@ func (s *SnowflakeSink) GetCopySink() *CopySink {
 
 // SnowflakeSource - A copy activity snowflake source.
 type SnowflakeSource struct {
+	// REQUIRED; Snowflake export settings.
+	ExportSettings *SnowflakeExportCopyCommand `json:"exportSettings,omitempty"`
+
 	// REQUIRED; Copy source type.
 	Type *string `json:"type,omitempty"`
 
@@ -26831,9 +27034,6 @@ type SnowflakeSource struct {
 
 	// If true, disable data store metrics collection. Default is false. Type: boolean (or Expression with resultType boolean).
 	DisableMetricsCollection interface{} `json:"disableMetricsCollection,omitempty"`
-
-	// Snowflake export settings.
-	ExportSettings *SnowflakeExportCopyCommand `json:"exportSettings,omitempty"`
 
 	// The maximum concurrent connection count for the source data store. Type: integer (or Expression with resultType integer).
 	MaxConcurrentConnections interface{} `json:"maxConcurrentConnections,omitempty"`